+++ RELEASES.md
@@ -43,6 +43,7 @@ Upcoming
 - Improve IBIS console function "Show Adapter Statistics"
 - Add attribute simple to ExchangeMailListener to save memory
 - Copy functionality to temporarily move and/or chomp received messages for memory purposes to pipes (next to receivers)
+- Add age of EMS server in GetTibcoQueues files (currently used in IJA_TiBeT2)